requirements.txt
openai>=1.0.0
anthropic>=0.7.0
requests>=2.28.0
python-dotenv>=0.19.0
pyyaml>=6.0
tqdm>=4.64.0
sqlite3
pandas>=1.5.0
tabulate>=0.9.0 